LayoutTests/fast/dom/StyleSheet/ownerNode-lifetime-2-expected.txt

CONSOLE MESSAGE: line 24: TypeError: Result of expression 'styleSheet' [null] is not an object.
This test verifies that style sheet do not outlive their elements. Since you can get back to the element with the ownerNode attribute, it's important to keep the element alive.

TEST HAS NOT RUN YET.
